### 2023-03-02 #5208 @editorialbot
[PRE REVIEW]: PreliZ: A tool-box for prior elicitation
**Submitting author:** @aleicazatti (<a href="http://orcid.org/0000-0003-1491-7330">Alejandro Icazatti</a>)
**Repository:** https://github.com/arviz-devs/preliz
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
