### 2023-03-02 3d185bd
Apply suggestions from JOSS editorial bot
* Add missing doi for sarma:2020
* Fix doi for morris:2014

https://github.com/openjournals/joss-reviews/issues/5208
