+++ paper/paper.bib
@@ -11,6 +11,7 @@
 
 @inproceedings{sarma:2020,
         title = {Prior {Setting} in {Practice}: {Strategies} and {Rationales} {Used} in {Choosing} {Prior} {Distributions} for {Bayesian} {Analysis}},
+        doi = {10.1145/3313831.3376377},
         booktitle = {Conference on {Human} {Factors} in {Computing} {Systems}},
         author = {Sarma, Abhraneel and Kay, Matthew},
         year = {2020},
@@ -57,7 +58,7 @@
         volume = {52},
         issn = {1364-8152},
         url = {https://www.sciencedirect.com/science/article/pii/S1364815213002533},
-        doi = {https://doi.org/10.1016/j.envsoft.2013.10.010},
+        doi = {10.1016/j.envsoft.2013.10.010},
         journal = {Environmental Modelling \& Software},
         author = {Morris, David E. and Oakley, Jeremy E. and Crowe, John A.},
         year = {2014},
