+++ preliz/distributions/discrete.py
@@ -848,7 +848,6 @@ class ZeroInflatedBinomial(Discrete):
         Number of Bernoulli trials (n >= 0).
     p : float
         Probability of success in each trial (0 < p < 1).
-
     """
 
     def __init__(self, psi=None, n=None, p=None):
@@ -889,6 +888,145 @@ class ZeroInflatedBinomial(Discrete):
         p = mean / n
         psi = 0.9
         params = psi, n, p
+        optimize_moments(self, mean, sigma, params)
+
+    def _fit_mle(self, sample):
+        optimize_ml(self, sample)
+
+
+class ZeroInflatedNegativeBinomial(Discrete):
+    R"""
+    Zero-Inflated Negative binomial log-likelihood.
+
+    The Zero-inflated version of the Negative Binomial (NB).
+    The NB distribution describes a Poisson random variable
+    whose rate parameter is gamma distributed.
+    The pmf of this distribution is
+
+    .. math::
+
+       f(x \mid \psi, \mu, \alpha) = \left\{
+         \begin{array}{l}
+           (1-\psi) + \psi \left (
+             \frac{\alpha}{\alpha+\mu}
+           \right) ^\alpha, \text{if } x = 0 \\
+           \psi \frac{\Gamma(x+\alpha)}{x! \Gamma(\alpha)} \left (
+             \frac{\alpha}{\mu+\alpha}
+           \right)^\alpha \left(
+             \frac{\mu}{\mu+\alpha}
+           \right)^x, \text{if } x=1,2,3,\ldots
+         \end{array}
+       \right.
+
+    .. plot::
+        :context: close-figs
+
+        import arviz as az
+        from preliz import ZeroInflatedNegativeBinomial
+        az.style.use('arviz-white')
+        psis = [0.7, 0.7]
+        mus = [2, 8]
+        alphas = [2, 4]
+        for psi, mu, alpha in zip(psis, mus, alphas):
+        ZeroInflatedNegativeBinomial(psi, mu=mu, alpha=alpha).plot_pdf(support=(0,25))
+
+    ========  ==========================
+    Support   :math:`x \in \mathbb{N}_0`
+    Mean      :math:`\psi\mu`
+    Var       :math:`\psi\mu +  \left (1 + \frac{\mu}{\alpha} + \frac{1-\psi}{\mu} \right)`
+    ========  ==========================
+
+    The zero inflated negative binomial distribution can be parametrized
+    either in terms of mu and alpha, or in terms of n and p.
+    The link between the parametrizations is given by
+
+    .. math::
+
+        \mu &= \frac{n(1-p)}{p} \\
+        \alpha &= n
+
+    Parameters
+    ----------
+    psi : float
+        Expected proportion of NegativeBinomial variates (0 < psi < 1)
+    mu : float
+        Poission distribution parameter (mu > 0).
+    alpha : float
+        Gamma distribution parameter (alpha > 0).
+    p : float
+        Alternative probability of success in each trial (0 < p < 1).
+    n : float
+        Alternative number of target success trials (n > 0)
+    """
+
+    def __init__(self, psi=None, mu=None, alpha=None, p=None, n=None):
+        super().__init__()
+        self.psi = psi
+        self.n = n
+        self.p = p
+        self.alpha = alpha
+        self.mu = mu
+        self.dist = ZINegativeBinomial
+        self.support = (0, np.inf)
+        self._parametrization(psi, mu, alpha, p, n)
+
+    def _parametrization(self, psi=None, mu=None, alpha=None, p=None, n=None):
+        if (mu or alpha) is not None and (p or n) is not None:
+            raise ValueError(
[…]
