### 2023-03-08 b90da3e
Add tests for ZeroInflatedNegativeBinomial
