### 2023-03-09 f995e76
Update preliz/distributions/discrete.py
