### 2023-03-09 a68f99b
Update preliz/distributions/discrete.py
