### 2023-03-09 e6bce34
Add ZeroInflatedNegativeBinomial (#222)
* [WIP] Add ZeroInflatedNegativeBinomial

* Add tests for ZeroInflatedNegativeBinomial

* Update preliz/distributions/discrete.py

* Update preliz/distributions/discrete.py

---------

Co-authored-by: Osvaldo A Martin <aloctavodia@gmail.com>
