+++ preliz/internal/distribution_helper.py
@@ -122,5 +122,6 @@ init_vals = {
     "NegativeBinomial": {"mu": 5.0, "alpha": 2.0},
     "Poisson": {"mu": 4.5},
     "ZeroInflatedBinomial": {"psi": 0.7, "n": 10, "p": 0.5},
+    "ZeroInflatedNegativeBinomial": {"psi": 0.7, "mu": 5, "alpha": 8},
     "ZeroInflatedPoisson": {"psi": 0.8, "mu": 4.5},
 }
