### 2023-03-10 48814aa
Update preliz/distributions/continuous_multivariate.py
Co-authored-by: Alejandro Icazatti <ale.icazatti@gmail.com>
