### 2023-03-10 e70551e
Update preliz/distributions/continuous_multivariate.py
Co-authored-by: Alejandro Icazatti <ale.icazatti@gmail.com>
