### 2023-03-10 39534e2
Apply suggestions from code review
Co-authored-by: Alejandro Icazatti <ale.icazatti@gmail.com>
