### 2023-03-10 a4d3dbd
Add initial support for MultiVariate Variables (#224)
* add multivariate distributions

* remove circular import

* Update preliz/distributions/continuous_multivariate.py

Co-authored-by: Alejandro Icazatti <ale.icazatti@gmail.com>

* Update preliz/distributions/continuous_multivariate.py

Co-authored-by: Alejandro Icazatti <ale.icazatti@gmail.com>

* Apply suggestions from code review

Co-authored-by: Alejandro Icazatti <ale.icazatti@gmail.com>

---------

Co-authored-by: Alejandro Icazatti <ale.icazatti@gmail.com>
