+++ preliz/distributions/distributions_multivariate.py
@@ -1,0 +1,220 @@
+"""
+Parent classes for multivariate families.
+"""
+# pylint: disable=no-member
+from collections import namedtuple
+
+import numpy as np
+from scipy.special import betainc  # pylint: disable=no-name-in-module
+
+from ..internal.distribution_helper import (
+    valid_scalar_params,
+    valid_distribution,
+)
+
+
+class Multivariate:
+    """
+    Base class for Multivariate distributions.
+
+    Not intended for direct instantiation.
+    """
+
+    def __init__(self):
+        self.rv_frozen = None
+        self.is_frozen = False
+        self.opt = None
+
+    def __repr__(self):
+        name = self.__class__.__name__
+        if self.is_frozen:
+            bolded_name = "\033[1m" + name + "\033[0m"
+
+            description = "".join(
+                f"{n}={np.round(v, 2)}," for n, v in zip(self.param_names, self.params)
+            ).strip(",")
+
+            return f"{bolded_name}({description})"
+        else:
+            return name
+
+    def _update_rv_frozen(self):
+        """Update the rv_frozen object"""
+
+        frozen = self._get_frozen()
+        if frozen is not None:
+            self.is_frozen = True
+            self.rv_frozen = frozen
+
+    def summary(self):
+        """
+        Namedtuple with the mean, and standard deviation of the distribution.
+        """
+        valid_distribution(self)
+
+        if valid_scalar_params(self):
+            attr = namedtuple(self.__class__.__name__, ["mean", "std"])
+            if self.__class__.__name__ == "MvNormal":
+                # for some weird reason, the mean of mvnormal is an attribute
+                # and not a method like for the rest of the distributions
+                mean = self.rv_frozen.mean
+            else:
+                mean = self.rv_frozen.mean()
+            std = self.rv_frozen.var() ** 0.5
+            return attr(mean, std)
+        else:
+            return None
+
+    def rvs(self, *args, **kwds):
+        """Random sample
+
+        Parameters
+        ----------
+        size : int or tuple of ints, optional
+            Defining number of random variates. Defaults to 1.
+        random_state : {None, int, numpy.random.Generator, numpy.random.RandomState}
+            Defaults to None
+        """
+        return self.rv_frozen.rvs(*args, **kwds)
+
+    def cdf(self, x):
+        """Cumulative distribution function.
+
+        Parameters
+        ----------
+        x : array_like
+            Values on which to evaluate the cdf
+        """
+        # not sure if this is right
+        return betainc(self.alpha, self.alpha.sum(), x)
+
+    def _check_endpoints(self, lower, upper, raise_error=True):
+        """
+        Evaluate if the lower and upper values are in the support of the distribution
+
+        Parameters
+        ----------
+        lower : int or float
+            lower endpoint
+        upper : int or float
+            upper endpoint
+        raise_error : bool
+            If True (default) it will raise ValueErrors, otherwise it will return True
+            if the lower and upper endpoint are in the support of the distribution or
+            False otherwise.
+        """
+        s_l, s_u = self.support
+
+        if raise_error:
+            domain_error = (
+                f"The provided endpoints are outside the domain of the "
+                f"{self.__class__.__name__} distribution"
+            )
+
+            if np.isfinite(s_l):
+                if lower < s_l:
+                    raise ValueError(domain_error)
+
+            if np.isfinite(s_u):
+                if upper > s_u:
+                    raise ValueError(domain_error)
+            if np.isfinite(s_l) and np.isfinite(s_u):
+                if lower == s_l and upper == s_u:
+                    raise ValueError(
+                        "Given the provided endpoints, mass will be always 1. "
+                        "Please provide other values"
[…]
