### 2023-03-11 d664c21
add continuous multivariate api reference (#228)
