### 2023-03-17 bfa0b2e
Add alternative parametrization for Wald
