+++ preliz/distributions/continuous.py
@@ -10,6 +10,7 @@ from copy import copy
 import numpy as np
 from scipy import stats
 from scipy.special import gamma as gammaf
+from scipy.special import beta as betaf
 from scipy.special import logit, expit  # pylint: disable=no-name-in-module
 
 from ..internal.optimization import optimize_ml, optimize_moments
@@ -1331,6 +1332,120 @@ class InverseGamma(Continuous):
     def _fit_mle(self, sample, **kwargs):
         alpha, _, beta = self.dist.fit(sample, **kwargs)
         self._update(alpha, beta)
+
+
+class Kumaraswamy(Continuous):
+    r"""
+    Kumaraswamy distribution.
+
+    The pdf of this distribution is
+
+    .. math::
+
+         f(x \mid a, b) = a b x^{a - 1} (1 - x^a)^{b - 1}
+
+    .. plot::
+        :context: close-figs
+
+        import arviz as az
+        from preliz import Kumaraswamy
+        az.style.use('arviz-white')
+        a_s = [.5, 5., 1., 2., 2.]
+        b_s = [.5, 1., 3., 2., 5.]
+        for a, b in zip(a_s, b_s):
+            ax = Kumaraswamy(a, b).plot_pdf()
+            ax.set_ylim(0, 3.)
+
+    ========  ==============================================================
+    Support   :math:`x \in (0, 1)`
+    Mean      :math:`b B(1 + \tfrac{1}{a}, b)`
+    Variance  :math:`b B(1 + \tfrac{2}{a}, b) - (b B(1 + \tfrac{1}{a}, b))^2`
+    ========  ==============================================================
+
+    Parameters
+    ----------
+    a : float
+        a > 0.
+    b : float
+        b > 0.
+    """
+
+    def __init__(self, a=None, b=None):
+        super().__init__()
+        self.dist = _Kumaraswamy
+        self.support = (0, 1)
+        self._parametrization(a, b)
+
+    def _parametrization(self, a=None, b=None):
+        self.a = a
+        self.b = b
+        self.params = (self.a, self.b)
+        self.param_names = ("a", "b")
+        self.params_support = ((0, np.inf), (0, np.inf))
+        if (a and b) is not None:
+            self._update(a, b)
+
+    def _get_frozen(self):
+        frozen = None
+        if all_not_none(self):
+            frozen = self.dist(self.a, self.b)
+        return frozen
+
+    def _update(self, a, b):
+        self.a = np.float64(a)
+        self.b = np.float64(b)
+        self.params = (self.a, self.b)
+        self._update_rv_frozen()
+
+    def _fit_moments(self, mean, sigma):
+        a = 1 / (1 - mean)
+        b = (1 - mean) / sigma**2
+        self._update(a, b)
+
+    def _fit_mle(self, sample, **kwargs):
+        a, b = self.dist.fit(sample, **kwargs)
+        self._update(a, b)
+
+
+class _Kumaraswamy(stats.rv_continuous):
+    def __init__(self, a=None, b=None):
+        super().__init__()
+        self.a = a
+        self.b = b
+
+    def support(self, *args, **kwargs):
+        return (0, 1)
+
+    def cdf(self, x, *args, **kwds):
+        return 1 - (1 - x**self.a) ** self.b
+
+    def pdf(self, x, *args, **kwds):
+        return (self.a * self.b * x ** (self.a - 1)) * ((1 - x**self.a) ** (self.b - 1))
+
+    def logpdf(self, x, *args, **kwds):
+        return (
+            np.log(self.a * self.b)
+            + (self.a - 1) * np.log(x)
[…]
