+++ preliz/tests/test_mle.py
@@ -18,6 +18,7 @@ from preliz.distributions import (
     HalfNormal,
     HalfStudentT,
     InverseGamma,
+    Kumaraswamy,
     Laplace,
     Logistic,
     LogNormal,
@@ -65,6 +66,7 @@ from preliz.distributions import (
         (HalfStudentT, (3, 1)),
         (HalfNormal, (2,)),
         (InverseGamma, (3, 5)),
+        (Kumaraswamy, (2, 3)),
         (Laplace, (0, 2)),
         (Logistic, (0, 1)),
         (LogNormal, (0, 1)),
