### 2023-03-27 81bcf9f
Add alternative parametrization for MvNormal
