### 2023-03-28 1a431e7
add any_not_none and all_not_none function
