+++ preliz/distributions/continuous.py
@@ -13,7 +13,7 @@ from scipy.special import gamma as gammaf
 from scipy.special import logit, expit  # pylint: disable=no-name-in-module
 
 from ..internal.optimization import optimize_ml, optimize_moments
-from ..internal.distribution_helper import garcia_approximation, all_not_none
+from ..internal.distribution_helper import garcia_approximation, all_not_none, any_not_none
 from .distributions import Continuous
 
 eps = np.finfo(float).eps
@@ -91,7 +91,7 @@ class AsymmetricLaplace(Continuous):
         self._parametrization(kappa, mu, b, q)
 
     def _parametrization(self, kappa=None, mu=None, b=None, q=None):
-        if kappa is not None and q is not None:
+        if all_not_none(kappa, q):
             raise ValueError("Incompatible parametrization. Either use kappa or q.")
 
         self.param_names = ("kappa", "mu", "b")
@@ -106,7 +106,7 @@ class AsymmetricLaplace(Continuous):
         self.kappa = kappa
         self.mu = mu
         self.b = b
-        if (kappa and mu and b) is not None:
+        if all_not_none(kappa, mu, b):
             self._update(kappa, mu, b)
 
     def _from_q(self, q):
@@ -119,7 +119,7 @@ class AsymmetricLaplace(Continuous):
 
     def _get_frozen(self):
         frozen = None
-        if all_not_none(self):
+        if all_not_none(self.params):
             frozen = self.dist(kappa=self.kappa, loc=self.mu, scale=self.b)
         return frozen
 
@@ -210,9 +210,9 @@ class Beta(Continuous):
 
     def _parametrization(self, alpha=None, beta=None, mu=None, sigma=None, kappa=None):
         if (
-            (alpha or beta) is not None
-            and (mu or sigma or kappa) is not None
-            or (sigma and kappa) is not None
+            any_not_none(alpha, beta)
+            and any_not_none(mu, sigma, kappa)
+            or all_not_none(sigma, kappa)
         ):
             raise ValueError(
                 "Incompatible parametrization. Either use alpha " "and beta, or mu and sigma."
@@ -221,25 +221,25 @@ class Beta(Continuous):
         self.param_names = ("alpha", "beta")
         self.params_support = ((eps, np.inf), (eps, np.inf))
 
-        if (mu or sigma) is not None:
+        if any_not_none(mu, sigma):
             self.mu = mu
             self.sigma = sigma
             self.param_names = ("mu", "sigma")
             self.params_support = ((eps, 1 - eps), (eps, 1 - eps))
-            if (mu and sigma) is not None:
+            if all_not_none(mu, sigma):
                 alpha, beta = self._from_mu_sigma(mu, sigma)
 
-        if (mu or kappa) is not None:
+        if any_not_none(mu, kappa):
             self.mu = mu
             self.kappa = kappa
             self.param_names = ("mu", "kappa")
             self.params_support = ((eps, 1 - eps), (eps, np.inf))
-            if (mu and kappa) is not None:
+            if all_not_none(mu, kappa):
                 alpha, beta = self._from_mu_kappa(mu, kappa)
 
         self.alpha = alpha
         self.beta = beta
-        if (self.alpha and self.beta) is not None:
+        if all_not_none(self.alpha, self.beta):
             self._update(self.alpha, self.beta)
 
     def _from_mu_sigma(self, mu, sigma):
@@ -261,7 +261,7 @@ class Beta(Continuous):
 
     def _get_frozen(self):
         frozen = None
-        if all_not_none(self):
+        if all_not_none(self.params):
             frozen = self.dist(self.alpha, self.beta)
         return frozen
 
@@ -347,12 +347,12 @@ class BetaScaled(Continuous):
[…]
