### 2023-03-28 8d372c6
add any_not_none and refactor all_not_none function (#233)
* add any_not_none and all_not_none function

* revert unintended change
