+++ preliz/distributions/discrete.py
@@ -17,7 +17,7 @@ from scipy.special import logit, expit  # pylint: disable=no-name-in-module
 
 from .distributions import Discrete
 from ..internal.optimization import optimize_ml, optimize_moments
-from ..internal.distribution_helper import all_not_none
+from ..internal.distribution_helper import all_not_none, any_not_none
 
 
 _log = logging.getLogger("preliz")
@@ -72,7 +72,7 @@ class Bernoulli(Discrete):
         self._parametrization(p, logit_p)
 
     def _parametrization(self, p=None, logit_p=None):
-        if p is not None and logit_p is not None:
+        if all_not_none(p, logit_p):
             raise ValueError("Incompatible parametrization. Either use p or logit_p.")
 
         self.param_names = "p"
@@ -95,7 +95,7 @@ class Bernoulli(Discrete):
 
     def _get_frozen(self):
         frozen = None
-        if all_not_none(self):
+        if all_not_none(self.params):
             frozen = self.dist(self.p)
         return frozen
 
@@ -174,12 +174,12 @@ class BetaBinomial(Discrete):
         self.params = (self.alpha, self.beta, self.n)
         self.param_names = ("alpha", "beta", "n")
         self.params_support = ((eps, np.inf), (eps, np.inf), (eps, np.inf))
-        if (alpha and beta) is not None:
+        if all_not_none(alpha, beta):
             self._update(alpha, beta, n)
 
     def _get_frozen(self):
         frozen = None
-        if all_not_none(self):
+        if all_not_none(self.params):
             frozen = self.dist(n=self.n, a=self.alpha, b=self.beta)
         return frozen
 
@@ -256,12 +256,12 @@ class Binomial(Discrete):
         self.params = (self.n, self.p)
         self.param_names = ("n", "p")
         self.params_support = ((eps, np.inf), (eps, 1 - eps))
-        if (n and p) is not None:
+        if all_not_none(n, p):
             self._update(n, p)
 
     def _get_frozen(self):
         frozen = None
-        if all_not_none(self):
+        if all_not_none(self.params):
             frozen = self.dist(self.n, self.p)
         return frozen
 
@@ -344,7 +344,7 @@ class Categorical(Discrete):
         return np.searchsorted(cumsum, q)
 
     def _parametrization(self, p=None, logit_p=None):
-        if p is not None and logit_p is not None:
+        if all_not_none(p, logit_p):
             raise ValueError("Incompatible parametrization. Either use p or logit_p.")
 
         self.param_names = "p"
@@ -368,7 +368,7 @@ class Categorical(Discrete):
 
     def _get_frozen(self):
         frozen = None
-        if all_not_none(self):
+        if all_not_none(self.params):
             frozen = self.dist(n=1, p=self.p)
         return frozen
 
@@ -439,7 +439,7 @@ class DiscreteUniform(Discrete):
         self.support = (self.lower, self.upper)
         self.dist.a = self.lower
         self.dist.b = self.upper
-        if (lower and upper) is not None:
+        if all_not_none(lower, upper):
             self._update(lower, upper)
         else:
             self.lower = lower
@@ -447,7 +447,7 @@ class DiscreteUniform(Discrete):
 
     def _get_frozen(self):
         frozen = None
-        if all_not_none(self):
+        if all_not_none(self.params):
             frozen = self.dist(self.lower, self.upper + 1)
         return frozen
 
@@ -517,7 +517,7 @@ class Geometric(Discrete):
 
     def _get_frozen(self):
         frozen = None
-        if all_not_none(self):
+        if all_not_none(self.params):
             frozen = self.dist(self.p)
         return frozen
 
[…]
