### 2023-03-28 3c72686
Add alternative parametrization for Wald (#230)
Co-authored-by: Osvaldo A Martin <aloctavodia@gmail.com>
