### 2023-03-29 37a310c
Apply suggestion from code review
Co-authored-by: Osvaldo A Martin <aloctavodia@gmail.com>
