### 2023-03-29 6253672
Add alternative parametrization for MvNormal (#232)
* Add alternative parametrization for MvNormal

* Apply suggestion from code review

Co-authored-by: Osvaldo A Martin <aloctavodia@gmail.com>

---------

Co-authored-by: Osvaldo A Martin <aloctavodia@gmail.com>
