### 2023-04-04 393bc1b
Add alternative parametrization for Exponential
