### 2023-04-06 33e9a74
Add Kumaraswamy distribution (#231)
* [WIP] Add Kumaraswamy distribution

* Complete Kumaraswamy distribution details

* Apply suggestions from code review

Co-authored-by: Osvaldo A Martin <aloctavodia@gmail.com>

* Update preliz/distributions/continuous.py

---------

Co-authored-by: Osvaldo A Martin <aloctavodia@gmail.com>
