+++ preliz/tests/test_quartile.py
@@ -17,6 +17,7 @@ from preliz.distributions import (
     HalfNormal,
     HalfStudentT,
     InverseGamma,
+    Kumaraswamy,
     Laplace,
     Logistic,
     LogNormal,
@@ -61,6 +62,7 @@ from preliz.distributions import (
         (HalfNormal(), 0.5, 1, 2, (1.613)),
         (HalfStudentT(), 0.5, 1, 2, (2.393, 1.311)),
         (InverseGamma(), 0.2, 0.3, 0.4, (3.881, 1.019)),
+        (Kumaraswamy(), 0.2, 0.3, 0.4, (2.199, 9.598)),
         (Laplace(), -1, 0, 1, (0, 1.442)),
         (Logistic(), -1, 0, 1, (0, 0.910)),
         (LogNormal(), 0.5, 1, 2, (0, 1.027)),
