### 2023-04-15 f173895
add alpha argument
