### 2023-04-17 3bae47e
faster kde predictive sliders  (#237)
* faster kde

* add alpha argument
