+++ preliz/__init__.py
@@ -14,7 +14,7 @@ from .unidimensional import *
 
 __all__ = ["maxent", "mle", "ppa", "roulette", "quartile"]
 
-__version__ = "0.2.0"
+__version__ = "0.3.0"
 
 _log = logging.getLogger("preliz")
 
