### 2023-04-30 e87376a
add color and alpha arguments to plot methods
