### 2023-05-03 7b6b2ae
Apply suggestion from code review
Co-authored-by: Osvaldo A Martin <aloctavodia@gmail.com>
