### 2023-05-04 e60e68e
add color and alpha arguments to plot methods (#240)
