### 2023-05-05 d49d721
[WIP] Add DiscreteWeibull  (#239)
* [WIP] Add DiscreteWeibull

* Fix some details

* Apply suggestion from code review

Co-authored-by: Osvaldo A Martin <aloctavodia@gmail.com>

* Black reformat

* Add init_vals in distribution_helper

---------

Co-authored-by: Osvaldo A Martin <aloctavodia@gmail.com>
