### 2023-05-07 78f8989
[WIP] Add tests for DiscreteWeibull
