+++ preliz/tests/test_distributions.py
@@ -39,6 +39,7 @@ from preliz.distributions import (
     BetaBinomial,
     Binomial,
     DiscreteUniform,
+    DiscreteWeibull,
     Geometric,
     NegativeBinomial,
     Poisson,
@@ -94,6 +95,7 @@ def multivariates():
         (Binomial, (2, 0.5)),
         (Binomial, (2, 0.1)),
         (DiscreteUniform, (0, 1)),
+        (DiscreteWeibull, (0.5, 3)),
         (Geometric, (0.75,)),
         (NegativeBinomial, (8, 4)),
         (Poisson, (4.5,)),
@@ -166,6 +168,7 @@ def test_moments(distribution, params):
         (Binomial, (2, 0.5)),
         (Binomial, (2, 0.1)),
         (DiscreteUniform, (0, 1)),
+        (DiscreteWeibull, (0.1, 1.5)),
         (Geometric, (0.5,)),
         (NegativeBinomial, (8, 4)),
         (Poisson, (4.5,)),
