+++ preliz/tests/test_maxent.py
@@ -40,6 +40,7 @@ from preliz.distributions import (
     BetaBinomial,
     Binomial,
     DiscreteUniform,
+    DiscreteWeibull,
     Geometric,
     HyperGeometric,
     NegativeBinomial,
@@ -145,6 +146,7 @@ from preliz.distributions import (
         (Binomial(), 3, 9, 0.9, (0, 9), (9, 0.490)),
         (Binomial(n=12), 3, 9, 0.9, (0, 12), (0.612)),
         (DiscreteUniform(), -2, 10, 0.9, (-3, 11), (-2, 10)),
+        (DiscreteWeibull(), 0, 6, 0.7, (0, 6), (0.85, 19.58)),
         (Geometric(), 1, 4, 0.99, (0, np.inf), (0.6837)),
         (HyperGeometric(), 2, 14, 0.9, (0, 21), (56, 21, 21)),
         (NegativeBinomial(), 0, 15, 0.9, (0, np.inf), (7.546, 2.041)),
