+++ preliz/tests/test_maxent.py
@@ -86,7 +86,7 @@ from preliz.distributions import (
         (HalfStudentT(), 1, 10, 0.7, (0, np.inf), (99.997, 7.697)),
         (HalfStudentT(nu=7), 1, 10, 0.7, (0, np.inf), (2.541)),
         (InverseGamma(), 0, 1, 0.99, (0, np.inf), (8.889, 3.439)),
-        (Kumaraswamy(), 0.1, 0.6, 0.9, (0, 1), (2.246, 7.426)),
+        (Kumaraswamy(), 0.1, 0.6, 0.9, (0, 1), (2.246, 7.427)),
         (Laplace(), -1, 1, 0.9, (-np.inf, np.inf), (0, 0.435)),
         (Laplace(mu=0.5), -1, 1, 0.9, (-np.inf, np.inf), (0.303)),
         (Logistic(), -1, 1, 0.5, (-np.inf, np.inf), (0, 0.91)),
@@ -101,7 +101,7 @@ from preliz.distributions import (
         (Normal(mu=0.5), -1, 1, 0.8, (-np.inf, np.inf), (0.581)),
         (Pareto(), 1, 4, 0.9, (1, np.inf), (1.660, 1)),
         (Pareto(m=2), 1, 4, 0.9, (2, np.inf), (3.321)),
-        (Rice(), 0, 4, 0.7, (0, np.inf), (0.0139, 2.577)),
+        (Rice(), 0, 4, 0.7, (0, np.inf), (0.0264, 2.577)),
         (Rice(nu=4), 0, 6, 0.9, (0, np.inf), (1.402)),
         (SkewNormal(), -2, 10, 0.9, (-np.inf, np.inf), (3.999, 3.647, 0)),
         (SkewNormal(mu=-1), -2, 10, 0.9, (-np.inf, np.inf), (6.2924, 4.905)),
