### 2023-05-08 71e9956
add figsize argument
