+++ preliz/tests/test_quartile.py
@@ -94,7 +94,7 @@ from preliz.distributions import (
         (HyperGeometric(), 2, 3, 4, (75, 15, 15)),
         (NegativeBinomial(), 3, 5, 10, (7.283, 2.167)),
         (Poisson(), 4, 5, 6, (5.641)),
-        (ZeroInflatedBinomial(), 3, 4, 7, (0.726, 17.001, 0.375)),
+        (ZeroInflatedBinomial(), 3, 4, 7, (0.720, 16.000, 0.400)),
         (ZeroInflatedBinomial(psi=0.7), 2, 4, 6, (16, 0.362)),
         (ZeroInflatedNegativeBinomial(), 2, 4, 6, (0.87, 5.24, 17.49)),
         (ZeroInflatedNegativeBinomial(psi=0.9), 2, 4, 6, (5.16, 11.32)),
