### 2023-05-08 c90ece3
Apply suggestions from code review
Co-authored-by: Osvaldo A Martin <aloctavodia@gmail.com>
