### 2023-05-09 fb7828d
add figsize argument (#244)
