### 2023-05-09 d138ad5
Rename fixed_lim to xy_lim
