### 2023-05-09 bf064e6
Rename `fixed_lim` to `xy_lim` (#245)
* Rename fixed_lim to xy_lim

* Fix typo
