### 2023-05-23 #5208 @djmannion
> @djmannion @GregoryAshton - would one of you be able to review this submission for JOSS, please?

Sure, I'd be happy to review it.
