### 2023-05-28 #5208 @olexandr-konovalov
@aleicazatti I have checked the latest article proof - before I will open the review for @djmannion and @jungtaekkim, would you be able to address some suggestions of mine, please?

- fix references (e.g. all of them in the "Statement of need" are not resolved,
- the reference to Morris et al. (page 2, line 46) does not have to be in quotes
- fix markdown (e.g. line 38 p.1) - instead of quotes, it should look like code in line 51, p.3
- same in line 44, page 2
- Fix figure captions (repeated word "Figure")
- check the use of punctuation in Acknowledgements

References:
- Fix the author of Ipywidgets reference.
- Also, add URL for their GitHub repository (is it https://github.com/jupyter-widgets/ipywidgets ?)
- Jupyter-matplotlib repository has a new URL: https://github.com/matplotlib/ipympl
