### 2023-05-29 ccc1b7b
convert input of cdf/pmf to array
