### 2023-05-29 53dc15b
make computation more robust
