+++ preliz/tests/test_maxent.py
@@ -102,7 +102,7 @@ from preliz.distributions import (
         (Normal(mu=0.5), -1, 1, 0.8, (-np.inf, np.inf), (0.581)),
         (Pareto(), 1, 4, 0.9, (1, np.inf), (1.660, 1)),
         (Pareto(m=2), 1, 4, 0.9, (2, np.inf), (3.321)),
-        (Rice(), 0, 4, 0.7, (0, np.inf), (0.0264, 2.577)),
+        (Rice(), 0, 4, 0.7, (0, np.inf), (0.017, 2.577)),
         (Rice(nu=4), 0, 6, 0.9, (0, np.inf), (1.402)),
         (SkewNormal(), -2, 10, 0.9, (-np.inf, np.inf), (3.999, 3.647, 0)),
         (SkewNormal(mu=-1), -2, 10, 0.9, (-np.inf, np.inf), (6.2924, 4.905)),
@@ -142,11 +142,11 @@ from preliz.distributions import (
         (Weibull(alpha=2), 0, 10, 0.9, (0, np.inf), (6.590)),
         (BetaBinomial(), 2, 8, 0.9, (0, 8), (1.951, 1.345, 8)),
         (BetaBinomial(n=10), 2, 6, 0.6, (0, 10), (1.837, 2.181)),
-        # results for binomial are close to the correct result, but still off
+        # # results for binomial are close to the correct result, but still off
         (Binomial(), 3, 9, 0.9, (0, 9), (9, 0.490)),
         (Binomial(n=12), 3, 9, 0.9, (0, 12), (0.612)),
         (DiscreteUniform(), -2, 10, 0.9, (-3, 11), (-2, 10)),
-        (DiscreteWeibull(), 0, 6, 0.7, (0, 6), (0.85, 19.58)),
+        (DiscreteWeibull(), 0, 6, 0.7, (0, np.inf), (0.85, 19.58)),
         (Geometric(), 1, 4, 0.99, (0, np.inf), (0.6837)),
         (HyperGeometric(), 2, 14, 0.9, (0, 21), (56, 21, 21)),
         (NegativeBinomial(), 0, 15, 0.9, (0, np.inf), (7.546, 2.041)),
