+++ paper/paper.md
@@ -71,6 +71,10 @@ PreliZ is written in Python 3.8+ and uses the following software packages:
 
 We thank our fiscal sponsor NumFOCUS, a nonprofit 501(c)(3) public charity, for their operational and financial support.
 
-This research was supported by the Academy of Finland Flagship Programme: Finnish Center for Artificial Intelligence FCAI. The National Agency of Scientific and Technological Promotion ANPCyT, Grant PICT-02212 (O.A.M.). And National Scientific and Technical Research Council CONICET, Grant PIP-0087 (O.A.M).
+This research was supported by:
+
+* the Academy of Finland Flagship Programme "Finnish Center for Artificial Intelligence" (FCAI)
+* the National Agency of Scientific and Technological Promotion (ANPCyT), Grant PICT-02212 (O.A.M.)
+* the National Scientific and Technical Research Council (CONICET), Grant PIP-0087 (O.A.M).
 
 # References
