### 2023-05-29 cb67d76
Apply suggestions from paper pre-review (#246)
* Apply suggestions from paper pre-review

* Improve acknowledgements
