+++ paper/paper.bib
@@ -69,7 +69,7 @@
 @misc{jupyter-matplotlib_2019,
         title = {jupyter-matplotlib},
         copyright = {BSD-3-Clause},
-        url = {https://github.com/matplotlib/jupyter-matplotlib},
+        url = {https://github.com/matplotlib/ipympl},
         abstract = {Matplotlib Jupyter Extension. Contribute to matplotlib/jupyter-matplotlib development by creating an account on GitHub.},
         urldate = {2019-11-19},
         publisher = {{Matplotlib Developers}},
@@ -79,10 +79,11 @@
 }
 
 @Unpublished{interactive_Jupyter_widgets,
-        author = {Jupyter widgets community},
+        author = {{Jupyter widgets community}},
         title = {ipywidgets, a GitHub repository},
         year = {2015},
-        note = {Retrieved from https://github.com/jupyter-widgets/ipywidgets}
+        url = {https://github.com/jupyter-widgets/ipywidgets},
+        note = {Retrieved from https://github.com/jupyter-widgets/ipywidgets},
 }
 
 @Article{Hunter:2007,
