### 2023-06-01 0d91c92
[WIP] Add tests for DiscreteWeibull (#242)
* [WIP] Add tests for DiscreteWeibull

* Apply suggestions from code review

* Apply suggestions from code review

Co-authored-by: Osvaldo A Martin <aloctavodia@gmail.com>

* convert input of cdf/pmf to array

* make computation more robust

* change entropy

---------

Co-authored-by: Osvaldo A Martin <aloctavodia@gmail.com>
