+++ preliz/tests/test_mle.py
@@ -39,6 +39,7 @@ from preliz.distributions import (
     BetaBinomial,
     Binomial,
     DiscreteUniform,
+    DiscreteWeibull,
     Geometric,
     HyperGeometric,
     NegativeBinomial,
@@ -87,6 +88,7 @@ from preliz.distributions import (
         (BetaBinomial, (1, 2, 10)),
         (Binomial, (5, 0.5)),
         (DiscreteUniform, (-2, 2)),
+        (DiscreteWeibull, (0.9, 1.3)),
         (Geometric, (0.75,)),
         (HyperGeometric, (50, 20, 10)),
         (NegativeBinomial, (10, 0.5)),
