### 2023-06-02 53337a6
use numerical approximation for entropy and continuous approximation for moments
