+++ preliz/distributions/discrete.py
@@ -12,7 +12,7 @@ from math import ceil
 
 import numpy as np
 from scipy import stats
-from scipy.special import logit, expit  # pylint: disable=no-name-in-module
+from scipy.special import logit, expit, gamma  # pylint: disable=no-name-in-module
 
 
 from .distributions import Discrete
@@ -516,7 +516,7 @@ class DiscreteWeibull(Discrete):
         self.beta = beta
         self.params = (self.q, self.beta)
         self.param_names = ("q", "beta")
-        self.params_support = ((eps, 1), (eps, np.inf))
+        self.params_support = ((eps, 1 - eps), (eps, np.inf))
         if all_not_none(q, beta):
             self._update(q, beta)
 
@@ -559,20 +559,38 @@ class _DiscreteWeibull(stats.rv_continuous):
 
     def logpmf(self, x, *args, **kwds):  # pylint: disable=unused-argument
         return np.log(self.pmf(x, *args, **kwds))
-    
+
     def ppf(self, p, *args, **kwds):  # pylint: disable=arguments-differ unused-argument
         p = np.asarray(p)
         p[p == 1] = 0.999999
-        return np.ceil(((np.log(1 - p) / np.log(self.q)) ** (1 / self.beta)) - 1)
+        ppf = np.ceil((np.log(1 - p) / np.log(self.q)) ** (1 / self.beta) - 1)
+        return ppf
 
     def _stats(self, *args, **kwds):  # pylint: disable=unused-argument
-        x_range =np.arange(1, np.nan_to_num(self._ppf(0.9999), nan=1)+1, dtype=int)
-        mean = np.sum(self.q ** (x_range**self.beta))
-        var = 2 * np.sum(x_range * self.q ** (x_range**self.beta)) - mean - mean**2
+        x_max = np.nan_to_num(self._ppf(0.999), nan=1)
+        if x_max < 10000:
+            x_range = np.arange(1, x_max + 1, dtype=int)
+            mean = np.sum(self.q ** (x_range**self.beta))
+            var = 2 * np.sum(x_range * self.q ** (x_range**self.beta)) - mean - mean**2
+        else:
+            lam = (-1 / np.log(self.q)) ** (1 / self.beta)
+            kappa = gamma(1 + 1 / self.beta)
+            mean = lam * kappa - 0.5
+            var = lam**2 * (gamma(1 + 2 / self.beta) - (kappa**2)) - 1
         return (mean, var, np.nan, np.nan)
 
     def entropy(self):  # pylint: disable=arguments-differ
-        return self.q / np.log(self.beta)
+        entropy = 0.0
+        x = 0
+        while True:
+            p_x = self.q ** (x**self.beta) - self.q ** ((x + 1) ** self.beta)
+            if p_x < 1e-6:
+                break
+            entropy -= p_x * np.log(p_x)
+            x += 1
+        return entropy
+
+        # return self.q / np.log(self.beta)
 
     def rvs(self, size=1, random_state=None):  # pylint: disable=arguments-differ
         return self.ppf(np.random.uniform(size=size), random_state=random_state)
