### 2023-06-08 6d2d7fe
use numerical approximation for entropy and continuous approximation for moments (#249)
