### 2023-06-08 cc79611
use koay method for estimation of Rice parameters
