### 2023-06-08 66a2a45
use eps instead of 0 for  Rayleigh approximation
