### 2023-06-09 0fe67fa
use koay method for estimation of Rice parameters (#251)
* use koay method for estimation of Rice parameters

* add initial guess

* use eps instead of 0 for  Rayleigh approximation
