+++ preliz/tests/test_maxent.py
@@ -102,7 +102,8 @@ from preliz.distributions import (
         (Normal(mu=0.5), -1, 1, 0.8, (-np.inf, np.inf), (0.581)),
         (Pareto(), 1, 4, 0.9, (1, np.inf), (1.660, 1)),
         (Pareto(m=2), 1, 4, 0.9, (2, np.inf), (3.321)),
-        # (Rice(), 0, 4, 0.7, (0, np.inf), (0.02647, 2.577)),
+        (Rice(), 0, 4, 0.7, (0, np.inf), (0, 2.577)),
+        (Rice(), 1, 10, 0.9, (0, np.inf), (3.453, 3.735)),
         (Rice(nu=4), 0, 6, 0.9, (0, np.inf), (1.402)),
         (SkewNormal(), -2, 10, 0.9, (-np.inf, np.inf), (3.999, 3.647, 0)),
         (SkewNormal(mu=-1), -2, 10, 0.9, (-np.inf, np.inf), (6.2924, 4.905)),
