### 2023-06-11 ea043c7
improve estimation sigma
