+++ preliz/tests/test_quartile.py
@@ -35,7 +35,7 @@ from preliz.distributions import (
     Weibull,
     # Bernoulli, quartile is not useful for Bernoulli distribution as we only have two states
     BetaBinomial,
-    # DiscreteUniform,
+    DiscreteUniform,
     DiscreteWeibull,
     Geometric,
     HyperGeometric,
@@ -89,15 +89,15 @@ from preliz.distributions import (
         (VonMises(), -1, 0, 1, (0, 0.656)),
         (Wald(), 0.5, 1, 2, (1.698, 1.109)),
         (Weibull(), 0.5, 1, 2, (1.109, 1.456)),
-        (BetaBinomial(), 2, 5, 8, (1.59, 4.49, 23)),
-        # (DiscreteUniform(), -1, 0, 1, (-5, 5)), # the mass is 0.27 instead of 0.5
+        (BetaBinomial(), 2, 5, 8, (1.182, 1.53, 13.0)),
+        (DiscreteUniform(), -2, 0, 2, (-5, 5)),
         (DiscreteWeibull(), 2, 6, 7, (0.951, 1.487)),
         (Geometric(), 2, 4, 6, (0.17)),
-        (HyperGeometric(), 2, 3, 4, (75, 15, 15)),
+        (HyperGeometric(), 2, 3, 4, (26, 8, 8)),
         (NegativeBinomial(), 3, 5, 10, (7.283, 2.167)),
         (Poisson(), 4, 5, 6, (5.641)),
-        (ZeroInflatedBinomial(), 3, 4, 7, (0.720, 16.000, 0.400)),
-        (ZeroInflatedBinomial(psi=0.7), 2, 4, 6, (16, 0.362)),
+        (ZeroInflatedBinomial(), 1, 4, 7, (0.683, 12, 0.540)),
+        (ZeroInflatedBinomial(psi=0.7), 2, 4, 6, (9, 0.6322)),
         (ZeroInflatedNegativeBinomial(), 2, 4, 6, (0.87, 5.24, 17.49)),
         (ZeroInflatedNegativeBinomial(psi=0.9), 2, 4, 6, (5.16, 11.32)),
         (ZeroInflatedPoisson(), 4, 5, 6, (1, 5.641)),
