+++ preliz/unidimensional/quartile.py
@@ -90,20 +90,16 @@ def quartile(
     # Heuristic to provide an initial guess for the optimization step
     # We obtain those guesses by first approximating the mean and standard deviation
     # from the quartiles and then use those values for moment matching
-    # distribution._fit_moments(mean=q2, sigma=(q3 - q1) * 1.5)  # pylint:disable=protected-access
     distribution._fit_moments(mean=q2, sigma=(q3 - q1) / 1.35)  # pylint:disable=protected-access
 
     opt = optimize_quartile(distribution, (q1, q2, q3), none_idx, fixed)
 
-    r_error, computed_mass = relative_error(distribution, q1, q3, 0.5)
+    r_error, _ = relative_error(distribution, q1, q3, 0.5)
 
     if r_error > 0.01:
         _log.info(
-            "The requested mass in the interval (q1=%.2g - q3=%.2g) is 0.5, "
-            "but the computed one is %.2g",
-            q1,
-            q3,
-            computed_mass,
+            "The expected masses are 0.25, 0.5, 0.75\n" "The computed ones are: %.2g, %.2g, %.2g",
+            *distribution.cdf([q1, q2, q3])
         )
 
     if plot:
