### 2023-06-11 arviz-devs/preliz#253 @djmannion
[paper] ppa model definition + typos
(I'm a reviewer for the submission of `preliz` at JOSS)

A few minor issues in the manuscript:
* Figure 4 (showing `pz.ppa`) refers to `a_preliz_model`, but that seems to be a different `a_preliz_model` than was defined in Figure 3. I suggest including the definition of `a_preliz_model` in Figure 4.
* Typo: "choosen" -> "chosen"
* Typo: "tipically" -> "typically"
