### 2023-06-11 arviz-devs/preliz#254 @djmannion
[docs] statement of need
(I'm a reviewer for the submission of `preliz` at JOSS)

This is in relation to the review criterion: "Documentation: A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?".

The initial landing page for the documentation (and also for the repo) could use an opening few sentences that meet this criterion (perhaps adapted from the paper).
