### 2023-06-12 arviz-devs/preliz#255 @djmannion
`notebook` dependency
(I'm a reviewer for the submission of `preliz` at JOSS)

This is in relation to the review criterion "Installation instructions: Is there a clearly-stated list of dependencies?".

Given the likely use of a jupyter notebook for the interactive aspects of the package, I'm wondering if `notebook` might be best listed as a dependency? Alternatively, perhaps it could be flagged in the "Installation" section of the docs that a `pip install notebook` may be required and again when `plot_interactive` is introduced.
