### 2023-06-12 #5499 @djmannion
Thanks to the authors for this very useful package! I have opened a few minor issues that would require consideration prior to the completion of my review:
* https://github.com/arviz-devs/preliz/issues/253
* https://github.com/arviz-devs/preliz/issues/254
* https://github.com/arviz-devs/preliz/issues/255
