### 2023-06-12 arviz-devs/preliz#254 @aloctavodia
Thanks, this is definitely something missing from the current docs.
