+++ docs/index.rst
@@ -1,8 +1,6 @@
 A tool-box for prior elicitation 
 ===================================================
 
-
-
 |PyPI version|
 |Tests|
 |Coverage|
@@ -19,6 +17,17 @@ A tool-box for prior elicitation
 
 .. |Black| image:: https://img.shields.io/badge/code%20style-black-000000.svg
     :target: https://github.com/ambv/black
+
+
+Overview
+========
+
+Prior elicitation refers to the process of transforming the knowledge of a particular domain into well-defined probability distributions.
+Specifying useful priors is a central aspect of Bayesian statistics. PreliZ is a Python package aimed at helping practitioners choose prior
+distributions by offering a set of tools for the various facets of prior elicitation. It covers a range of methods, from unidimensional prior
+elicitation on the parameter space to predictive elicitation on the observed space. The goal is to be compatible with probabilistic programming
+languages (PPL) in the Python ecosystem like PyMC and PyStan, while remaining agnostic of any specific PPL.
+
 
 Dependencies
 ============
