### 2023-06-12 ede0fea
Quartile init value and message (#256)
* improve estimation sigma

* report estimated masses
