### 2023-06-12 arviz-devs/preliz#255 @aloctavodia
Thanks for the feedback. @OriolAbril  @aleicazatti which one of the options do you like best?
