+++ docs/index.rst
@@ -36,24 +36,25 @@ PreliZ is tested on Python 3.8+.
 Installation
 ============
 
-Two dependency bundles are available for the latest release:
+For the latest release (base set of dependencies) you can do:
 
-* ``full``: includes dependencies for all features (interactive and non-interactive)
-* ``non-interactive``: includes dependencies for non-interactive features only.
+.. code-block:: bash
 
-You can install them with:
+  pip install preliz
+
+To make use of the interactive features, you can install the optional dependencies:
 
 .. tabs::
 
-  .. tab:: Full
+  .. tab:: JupyterLab
     .. code-block:: bash
 
-      pip install "preliz[full]"
+      pip install "preliz[full,lab]"
 
-  .. tab:: Non-interactive
+  .. tab:: Jupyter Notebook
     .. code-block:: bash
 
-      pip install "preliz[non-interactive]"
+      pip install "preliz[full,notebook]"
 
 The latest development version can be installed from the main branch using pip:
 
