+++ pyproject.toml
@@ -26,10 +26,6 @@ dependencies = [
   "arviz",
   "numpy>=1.22",
   "scipy>=1.9.1",
-  "matplotlib>=3.5",
-  "nbclient<0.6,>=0.2",
-  "ipywidgets",
-  "ipympl",
 ]
 
 [tool.flit.module]
@@ -41,21 +37,13 @@ tracker = "https://github.com/arviz-devs/preliz/issues"
 
 [project.optional-dependencies]
 full = [
-  "arviz",
-  "numpy>=1.22",
-  "scipy>=1.9.1",
   "matplotlib>=3.5",
   "nbclient<0.6,>=0.2",
   "ipywidgets",
   "ipympl",
-  "jupyterlab",
-  "notebook",
 ]
-non-interactive = [
-  "arviz",
-  "numpy>=1.22",
-  "scipy>=1.9.1",
-]
+lab = ["jupyterlab"]
+notebook = ["notebook"]
 
 [tool.black]
 line-length = 100
