### 2023-06-15 83d8204
install [full] for docs and tests
