### 2023-06-15 e3af450
Update .readthedocs.yaml: doc -> full
Don't know what I was thinking with doc
