+++ .readthedocs.yaml
@@ -13,4 +13,4 @@ python:
    - method: pip
      path: .
      extra_requirements:
-       - doc
+       - full
