### 2023-06-17 arviz-devs/preliz#260 @jungtaekkim
JOSS reviews for a paper
Hi the authors of PreliZ,

I would like to suggest a few revisions for your paper.

I think that the style of section titles should be consistent.  Please choose title case or sentence case and keep consistency.

Codes in figures should be written in texts, instead of images.  Please use code blocks in Markdown.

DOIs of some references are missing.  `Bayesian Modeling and Computation in Python` might be https://doi.org/10.1201/9781003019169.

Best regards,
Jungtaek.
