### 2023-06-21 03f0934
Add matplotlib to base dependencies
