### 2023-06-21 7053515
Add dependency bundles (#259)
* Add dependency bundles

* Update requirements-docs

* Apply suggestions from code review

* install [full] for docs and tests

* Update .readthedocs.yaml: doc -> full

Don't know what I was thinking with doc

* Add matplotlib to base dependencies

---------

Co-authored-by: Oriol Abril-Pla <oriol.abril.pla@gmail.com>
