+++ .readthedocs.yaml
@@ -3,12 +3,14 @@
 version: 2
 
 build:
-  os: ubuntu-20.04
+  os: ubuntu-22.04
   tools:
-    python: "3.9"
+    python: "3.10"
 
 python:
    install:
    - requirements: requirements-docs.txt
    - method: pip
      path: .
+     extra_requirements:
+       - full
