+++ paper/paper.bib
@@ -19,6 +19,8 @@
 }
 
 @book{martin:2021,
+        doi = {10.1201/9781003019169}, 
+        url = {https://doi.org/10.1201/9781003019169}, 
         address = {Boca Raton},
         title = {Bayesian {Modeling} and {Computation} in {Python}},
         isbn = {978-0-367-89436-8},
