+++ paper/paper.bib
@@ -45,6 +45,8 @@
 }
 
 @book{gelman:2013,
+        url = {https://doi.org/10.1201/b16018},
+        doi = {10.1201/b16018},
         address = {Boca Raton},
         edition = {3 edition},
         title = {Bayesian {Data} {Analysis}, {Third} {Edition}},
@@ -148,9 +150,14 @@
         doi     = {10.1038/s41592-019-0686-2},
 }
 
-@article{mikkola:2021,
-        title={Prior knowledge elicitation: The past, present, and future},
-        author={Mikkola, Petrus and Martin, Osvaldo A and Chandramouli, Suyog and Hartmann, Marcelo and Pla, Oriol Abril and Thomas, Owen and Pesonen, Henri and Corander, Jukka and Vehtari, Aki and Kaski, Samuel and others},
-        journal={arXiv preprint arXiv:2112.01380},
-        year={2021}
+@Article{mikkola:2021,
+        author = {Petrus Mikkola and Osvaldo A. Martin and Suyog Chandramouli and Marcelo Hartmann and Oriol Abril Pla and Owen Thomas and Henri Pesonen and Jukka Corander and Aki Vehtari and Samuel Kaski and Paul-Christian B{"u}rkner and Arto Klami},
+        title = {{Prior Knowledge Elicitation: The Past, Present, and Future}},
+        journal = {Bayesian Analysis},
+        publisher = {International Society for Bayesian Analysis},
+        pages = {1 -- 33},
+        keywords = {Bayesian workflow, domain knowledge, informative prior, prior distribution, prior elicitation},
+        year = {2023},
+        doi = {10.1214/23-BA1381},
+        URL = {https://doi.org/10.1214/23-BA1381},
 }
