### 2023-06-22 6d911da
Apply suggestions from paper review (#263)
* Apply suggestions from paper review

* Update references
