+++ paper/paper.md
@@ -37,26 +37,55 @@ In a Bayesian modeling workflow, a prior distribution can be chosen in different
 
 Specifying useful priors is a central aspect of Bayesian statistics [@gelman:2013; @martin:2021; @martin:2022], yet prior elicitation techniques that would make this step easier and more systematic are not routinely used within practical Bayesian workflows [@mikkola:2021]. Instead, practitioners tipicaly rely on ad hoc procedures based on a mix of their own experience and recommendations available in the literature, which in general has not been systematized [@sarma:2020]. One reason is that current solutions are simply not sufficient for practical data analysis and do not integrate well with probabilistic programming libraries [@mikkola:2021]. PreliZ is a library for prior elicitation that aims to facilitate the task for practitioners by offering a set of tools for the various facets of prior elicitation. It covers a range of methods, from unidimensional prior elicitation on the parameter space to predictive elicitation on the observed space. The goal is to be compatible with probabilistic programming languages in the Python ecosystem like PyMC and PyStan, which involves including the distributions and their parameterization that are supported by these languages.
 
-# Software API and Features
+# Software API and features
 
 PreliZ provides functionality for four core aspects: (1) Specification and manipulation of prior distributions, (2) Visualization of priors and the induced predictive distributions, (3) User interaction components, and (4) Algorithms learning priors from user-provided information.
 PreliZ distributions are built on top of SciPy distributions, providing easy access to all standard functionality, in particular random sampling and quantiles needed in the elicitation process. However, PreliZ extends many of the distributions for increased user-friendliness. For instance, PreliZ uses a canonical parametrization, instead of loc-scale style used by many of the SciPy distributions, and adds common alternative parametrizations,  e.g. Beta-distribution also in terms of mean and standard deviation. Finally, PreliZ also includes some distributions not available in SciPy, like the HalfStudentT, or LogitNormal.
 
 Preliz provides a range of plotting commands, including probability densities and cumulative densities (`plot_pdf` and `plot_cdf`) as well as extensive numerical summaries covering e.g. equal tailed intervals and highest density intervals. All plots support easy integration of input functionality, with the method `plot_interactive` automatically creating ipywidgets sliders for effortless exploration of how the parameters affect the distribution (see \autoref{fig:figure_1}).
 
+```python
+pz.Gamma(mu=2, sigma=1).plot_interactive()
+```
+
 ![Example of `plot_interactive()` output for Gamma distribution with mu=2 and sigma=1.\label{fig:figure_1}](figures/figure_1.png)
 
 PreliZ includes a few algorithms for inferring the suitable prior from the user-provided information, and provides a general API that makes adding new algorithms easy. As a practical example, `pz.maxent(pz.Gamma(mu=4), 1, 10, 0.9)` would find the maximum-entropy Gamma distribution that has 90% of its mass between 1 and 10, with the additional constraint of mean being 4. The library currently implements also the roulette method of @morris:2014, allowing users to find a prior distribution by drawing a histogram (see \autoref{fig:figure_2}).
+
+```python
+%matplotlib widget
+pz.roulette()
+```
 
 ![Roulette method.\label{fig:figure_2}](figures/figure_2.png)
 
[…]
