### 2023-06-29 #5499 @jungtaekkim
I reviewed this paper and the issue raised was resolved.

- https://github.com/arviz-devs/preliz/issues/260

I think this paper is ready to publish.
