### 2023-07-07 f5a5c28
pip scipy version
