+++ pyproject.toml
@@ -26,7 +26,7 @@ dependencies = [
   "arviz",
   "matplotlib>=3.5",
   "numpy>=1.22",
-  "scipy>=1.9.1",
+  "scipy>=1.9.1,<1.11.0",
 ]
 
 [tool.flit.module]
