### 2023-07-31 fc45d13
add try-except (#268)
