+++ .github/workflows/test.yml
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.9", "3.10", 3.11"]
+        python-version: ["3.9", "3.10", "3.11"]
 
     name: Set up Python ${{ matrix.python-version }}
     steps:
