### 2023-08-02 1a2db2f
test 3.11 (#270)
* test 3.11

* fix string

* make pylint happy

* make pylint happy

* try ipykernel
