+++ preliz/__init__.py
@@ -14,7 +14,7 @@ from .unidimensional import *
 
 __all__ = ["maxent", "mle", "ppa", "roulette", "quartile"]
 
-__version__ = "0.3.1"
+__version__ = "0.3.2"
 
 _log = logging.getLogger("preliz")
 
