### 2023-08-11 81f3a8c
change test trunctatednormal
