### 2023-08-11 01986c0
fix for scipy 1.11.1 (#272)
* fix for scipy 1.11.1

* change test trunctatednormal
