### 2023-08-29 b997441
allow singular covariance (#274)
