### 2023-09-21 #5499 @olexandr-konovalov
@aloctavodia @aleicazatti sorry it took a while to get back to you. I've made another check of the PDF:
- line 43 of `paper.md` - missing "to" in "helping practitioners choose prior distributions"
- bibliography: the new name of `jupyter-matplotlib` is `ipympl` https://matplotlib.org/ipympl/
- page 4 only has a figure and a code snippet. Perhaps some reaarangement or adding some textual "glue" might help to avoid that. Otherwise, it's quite unclear what the code belongs to.

Could you please fix that and I will then recommend it for acceptance.
