### 2023-09-21 8b4a1db
update following reviewer suggestions (#275)
