+++ paper/paper.bib
@@ -71,7 +71,7 @@
 }
 
 @misc{jupyter-matplotlib_2019,
-        title = {jupyter-matplotlib},
+        title = {ipympl},
         copyright = {BSD-3-Clause},
         url = {https://github.com/matplotlib/ipympl},
         abstract = {Matplotlib Jupyter Extension. Contribute to matplotlib/jupyter-matplotlib development by creating an account on GitHub.},
