### 2023-09-21 60a26f2
Fix LaTeX umlaut
