+++ paper/paper.bib
@@ -151,7 +151,7 @@
 }
 
 @Article{mikkola:2021,
-        author = {Petrus Mikkola and Osvaldo A. Martin and Suyog Chandramouli and Marcelo Hartmann and Oriol Abril Pla and Owen Thomas and Henri Pesonen and Jukka Corander and Aki Vehtari and Samuel Kaski and Paul-Christian B{"u}rkner and Arto Klami},
+        author = {Petrus Mikkola and Osvaldo A. Martin and Suyog Chandramouli and Marcelo Hartmann and Oriol Abril Pla and Owen Thomas and Henri Pesonen and Jukka Corander and Aki Vehtari and Samuel Kaski and Paul-Christian B{\"u}rkner and Arto Klami},
         title = {{Prior Knowledge Elicitation: The Past, Present, and Future}},
         journal = {Bayesian Analysis},
         publisher = {International Society for Bayesian Analysis},
