### 2023-09-21 #5499 @olexandr-konovalov
@aloctavodia thanks - please tick all what you've checked :) Also see https://github.com/arviz-devs/preliz/pull/277
