### 2023-09-21 #5499 @olexandr-konovalov
@aloctavodia also, please edit the title at https://zenodo.org/record/8368036: it should be

PreliZ: A tool-box for prior elicitation

instead of

arviz-devs/preliz: 0.3.3
