### 2023-09-21 #5499 @olexandr-konovalov
P.S. @aloctavodia I think the correct order would be to
- merge https://github.com/arviz-devs/preliz/pull/277
- make new release to keep in sync with the repository
- let me know new version number and DOI

It seems like everything else is in a good shape, so hopefully no more quick releases will be needed (until it will go to the EiC).
