### 2023-09-21 #5499 @aloctavodia
I tick them, but they get unticked when I refresh the page. I tried from my laptop and from my mobile phone.
