+++ preliz/__init__.py
@@ -14,7 +14,7 @@ from .unidimensional import *
 
 __all__ = ["maxent", "mle", "ppa", "roulette", "quartile"]
 
-__version__ = "0.3.3"
+__version__ = "0.3.4"
 
 _log = logging.getLogger("preliz")
 
