### 2023-09-21 #5499 @olexandr-konovalov
Thanks @aloctavodia! Looking at https://zenodo.org/record/8368516, you can also edit metadata, so that it will have some short abstract instead of "What's Changed" and "New Contributors" - this version will be linked from the article, so it would be more informative and appropriate.
