### 2023-09-22 973d05e
add citation (#279)
